run --ov --samefs uses lower/upper on same fs
When using flag --samefs, /lower and /upper are
bind mounts to /base/lower base/upper on the same fs.

If an entry for /base exists in fstab, e.g.:
 /dev/lower_layer           /base xfs noauto
it will be mounted as base fs.
Otherwise, /base will be tmpfs.

It is essential for the name of the base block
device to be 'lower_layer'.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>