Verify constant inode number for non same fs
Directory inode number for non same fs should be constant, but not
persistent, so check only for no mount cycle test.

The test 'run --ov --verify --samefs' is expected to fail the
constant inode check on kernels < v4.12, due to missing commit
72b608f08528 ("ovl: constant st_ino/st_dev across copy up").

The test 'run --ov=1 --verify --samefs' is expected to fail the
persistent inode check on kernels < v4.13, due to missing commit
2b4db79618ad ("tmpfs: generate random sb->s_uuid").

The test 'run --ov --verify' (non samefs) is expected to fail the
constant inode check on kernels <= v4.13.
A fix for constant inode number for non samefs is in the works.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>