Update layer check for non samefs --verify case
With samefs setup, files are on overlay st_dev if st_ino is constant
on copy up and on real st_dev if st_ino is not constant.
--verify verifies constant st_ino, so it implies overlay st_dev check.
Without --verify, we allow for both options.

With non samefs setup, files are on pseudo or upper st_dev if st_ino
is constant on copy up and on lower or upper st_dev otherwise.
--verify verifies constant st_ino, so it implies pseudo st_dev check.
Without --verify we allow for both options.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>