Add option run --ov=N --maxfs=M to configure number of different fs
Until now we had two configurations: samefs and not samefs.
--samefs used samefs for /upper and /lower and not samefs (default)
used one fs for /upper and one fs for /lower.

With the option --ov=N upper layer may be rotated up to N times
and create new layers at /upper/N, but until now, all layers
/upper/N where on the same fs.

With --maxfs=M it is possible to configure that new upper layers
are created as unique tmpfs:

 /upper/N are all same fs (/upper) if maxfs == 0
 /upper/N where N < maxfs are unique fs
 /upper/N where N >= maxfs are same fs (/upper)

Signed-off-by: Amir Goldstein <amir73il@gmail.com>