Record meta copy_up vs. data copy_up
Convert the boolean __upper to an IntEnum that can take
None, upper.META, upper.DATA and record the expected
result with metacopy=on. Add a method data_on_upper() to
check if data is expected to be copied up.

Originally, the on_upper() method was used in check_layer()
to expect st_dev to be from lower or upper dev, but nowadays,
in most configurations, overlayfs st_dev is the overlay dev.

The only case where st_dev can be the upper fs dev is for
pure upper file, so covert the check to use data_on_upper().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>