Don't copy files to snasphot with --meta
When running snapshot test with --meta, disable copy up
of files (copy only directories).

With --meta --verify, check that upper layers of metacopy
snapshots do not contain any regular files. When comparing
metacopy snapshot to backup with --verify, allow files to
differ from metacopy snapshot to backup and to be only in backup,
but they cannot be only in snapshot because newly created
files/dirs should be whited out in metacopy snapshot.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>