Verify integrity of all snapshots with ./run --sn --verify
Make full backup copy of lower layer on every rotate upper.
The location of the full backups is /base/b/{0..N}, where
/base/b/0 is the backup that was created during test setup.

At the end of the test run, compare all the snapshots to all
the full backups.

Allow files only in backup, but not only in snapshot, because
rename of directory can look like recursive remove in snapshot.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>