Test mkdir under renamed directory and rename it back
When creating a directory in a subdir of renamed directory, a whiteout
is not created in snapshot, because renamed directory is already a
whiteout.

When directory is renamed back, if upper dir above it is not set opaque
the created subdir will be exposed in snapshot.

This is observed with command line:
./run --sn --verify --recycle rename-pop-dir

--recycle is needed to reproduce the bug, because without it, the upper
dir can still be found in snapshot overlay after rename by inode cache
and grandchild whiteout is created.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>