Relax layers check for older kernels
The layers check was updated recently to check the upstream
behavior that exposes a pseudo dev for st_dev of overlay objects.
To support running the test on older kernels, before lower objects
exposed a pseudo st_dev, do not fail the layers check when detecting
an object with lower fs st_dev, unless --verify option was specified.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>