Auto enable xino also on kernel < v5.7
The meaning of "xino=auto" option has changed since kernel commit
926e94d79baf ("ovl: enable xino automatically in more cases") in
kernel v5.7 and it is now logically the same as option "xino=on"
except for verbosity level of warnings on inode number overflow.

Conform to the same behavior when testing kernel < v5.7 and promote
the option "xino=auto" to "xino=on". We do not care to get extra
warnings while testing - on the contrary.

Reported-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Tested-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>