Test concurrent overlay snapshot watches
Keep old snapshot watches mounted unless mount cycle was requested.
Changes are recorded to all old snapshots, so they are differential
to first backup.

Old snapshot fs implementation does not support concurrent snapshots
so in that case and in case of --recycle, we need to stack the old
incremental snapshots as lower layers to get the diff from first backup.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>