Test new overlay snapshot watch
The overlay mount with change snapshot watch does not need the
snapshot fs mount, so it is replaced with a bind mount from lower to
union mnt.

We only support "meta" snapshot now, so always use meta snapshot
verification and re-use the old --meta flag to request using the
snapshot fs implemenetation on old kernels.

On new kernels with no snapshot fs support, --meta is ignored.
The --remount flag is also ignored for snapshot fs tests.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>