Fail overlay watch verification if no changes were recorded
Verification uses du -x to list the directories recorded as changed
in snapshot, but that only works correctly when starting du -x on
a changed directory.

The verification code was running 'du -x a' but if 'a' was not recorded
as changed in snapshot, all non changed directories would be listed as
changed.

Change the code to run 'du -x .' and filter out the unneeded './'
from output.

This was tested by changing the overlay watch mount option to
watch=mnt which results in no changed being recorded due to
running the test cases on a bind mount of lower.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>